requirements.txt
gradio
torch
numpy
opencv-python
Pillow
diffusers
transformers
accelerate